+++ library/src/main/java/org/mustangproject/ZUGFeRD/IZUGFeRDExportableItem.java
@@ -26,11 +26,14 @@ package org.mustangproject.ZUGFeRD;
  * @author jstaerk
  * */
 
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.mustangproject.Item;
+
 import java.math.BigDecimal;
 import java.util.Date;
 
+@JsonDeserialize(as = Item.class)
 public interface IZUGFeRDExportableItem extends IAbsoluteValueProvider{
-
 	IZUGFeRDExportableProduct getProduct();
 
 	/**